feat: add "classic" version of the plugin
