+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "latency-slider-de"
-version = "0.3.2"
+version = "0.4.0"
 authors = []
 edition = "2021"
 
@@ -14,7 +14,11 @@ crate-type = ["cdylib"]
 ninput = { git = "https://github.com/blu-dev/ninput", version = "0.1.0" }
 skyline_smash = { git = "https://github.com/jugeeya/skyline-smash.git", branch = "patch-2" }
 skyline = { git = "https://github.com/ultimate-research/skyline-rs.git" }
-once_cell = { version = "1.20.2" }
+once_cell = { version = "1.20.2", default-features = true }
+
+[features]
+default = ["css_ui"]
+css_ui = []
 
 [profile.dev]
 panic = "abort"
