+++ src/offsets.rs
@@ -82,17 +82,6 @@ impl SSBUMemoryLocation<'_> {
     }
 }
 
-pub static LOC_DRAW: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0x08, 0x0c, 0x40, 0xf9, 0xc8, 0x03, 0x00, 0xb4, 0xff, 0x83, 0x01, 0xd1, 0xf5, 0x1b, 0x00,
-        0xf9, 0xf4, 0x4f, 0x04, 0xa9, 0xfd, 0x7b, 0x05, 0xa9, 0xfd, 0x43, 0x01, 0x91, 0xf4, 0x03,
-        0x00, 0xaa,
-    ],
-    start_offset: 0,
-    location_name: "draw",
-    cached_offset: OnceCell::new(),
-};
-
 pub static LOC_SET_ONLINE_LATENCY: SSBUMemoryLocation = SSBUMemoryLocation {
     signature: &[
         0xfd, 0x7b, 0x42, 0xa9, 0xf4, 0x4f, 0x41, 0xa9, 0xe8, 0x07, 0x43, 0xfc, 0xc0, 0x03, 0x5f,
@@ -100,68 +89,6 @@ pub static LOC_SET_ONLINE_LATENCY: SSBUMemoryLocation = SSBUMemoryLocation {
     ],
     start_offset: 0,
     location_name: "set_online_latency",
-    cached_offset: OnceCell::new(),
-};
-
-pub static LOC_INGAME_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0xff, 0x83, 0x01, 0xd1, 0xfa, 0x67, 0x01, 0xa9, 0xf8, 0x5f, 0x02, 0xa9, 0xf6, 0x57, 0x03,
-        0xa9, 0xf4, 0x4f, 0x04, 0xa9, 0xfd, 0x7b, 0x05, 0xa9, 0xfd, 0x43, 0x01, 0x91, 0xf5, 0x03,
-        0x01, 0xaa, 0xf3, 0x03, 0x00, 0xaa, 0x16, 0x40, 0x01, 0x91, 0x01, 0x37, 0x80, 0x52, 0xe0,
-        0x03, 0x1c, 0x32,
-    ],
-    start_offset: 48,
-    location_name: "ingame_scene",
-    cached_offset: OnceCell::new(),
-};
-
-pub static LOC_ONLINE_MELEE_ANY_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0xe8, 0x69, 0x01, 0xb0, 0x08, 0x01, 0x00, 0x91, 0x1f, 0x40, 0x00, 0x39, 0x08, 0x7c, 0x00,
-        0xa9, 0x1f, 0x40, 0x01, 0x39, 0x1f, 0x54, 0x00, 0xb9, 0x1f, 0x2c, 0x00, 0xf9, 0xfd, 0x7b,
-        0x41, 0xa9, 0xff, 0x83, 0x00, 0x91, 0xc0, 0x03, 0x5f, 0xd6,
-    ],
-    start_offset: 12,
-    location_name: "online_melee_any_scene",
-    cached_offset: OnceCell::new(),
-};
-
-pub static LOC_MAIN_MENU_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0x48, 0x6a, 0x01, 0xb0, 0x08, 0xe1, 0x18, 0x91, 0x7f, 0x42, 0x00, 0x39, 0x68, 0x7e, 0x00,
-        0xa9, 0xa8, 0x7e, 0x01, 0xb0, 0x7f, 0x42, 0x01, 0x39, 0xe1, 0x03, 0x00, 0x32, 0x7f, 0x56,
-        0x00, 0xb9, 0x7f, 0x2e, 0x00, 0xf9, 0x00, 0xf9, 0x43, 0xf9, 0xc7, 0x43, 0x3e, 0x94, 0xe0,
-        0x03, 0x13, 0xaa, 0xfd, 0x7b, 0x42, 0xa9, 0xf3, 0x0b, 0x40, 0xf9, 0xff, 0xc3, 0x00, 0x91,
-        0xc0, 0x03, 0x5f, 0xd6,
-    ],
-    start_offset: 12,
-    location_name: "main_menu_scene",
-    cached_offset: OnceCell::new(),
-};
-
-pub static LOC_MELEE_NORMAL_SEQUENCE_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0xc8, 0x6e, 0x01, 0xf0, 0x08, 0xc1, 0x0b, 0x91, 0x60, 0x42, 0x01, 0x91, 0x7f, 0x42, 0x00,
-        0x39, 0x68, 0x7e, 0x00, 0xa9, 0x64, 0x1f, 0x51, 0x94, 0x28, 0x6e, 0x01, 0xd0, 0x08, 0xc1,
-        0x01, 0x91, 0xe9, 0x6e, 0x01, 0x90, 0x29, 0xe1, 0x29, 0x91, 0x7f, 0xc2, 0x01, 0x79, 0x7f,
-        0xe6, 0x00, 0xb9, 0x68, 0x2a, 0x00, 0xf9, 0x69, 0x02, 0x00, 0xf9, 0xe0, 0x03, 0x13, 0xaa,
-        0x7f, 0x76, 0x00, 0xf9, 0xfd, 0x7b, 0x42, 0xa9, 0xf3, 0x0b, 0x40, 0xf9, 0xff, 0xc3, 0x00,
-        0x91, 0xc0, 0x03, 0x5f, 0xd6,
-    ],
-    start_offset: 20,
-    location_name: "melee_normal_sequence_scene",
-    cached_offset: OnceCell::new(),
-};
-
-pub static LOC_DISPLAY_CSS: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0xa0, 0x83, 0x53, 0xf8, 0x34, 0x00, 0x00, 0x94, 0xff, 0x03, 0x0e, 0x91, 0xe9, 0x23, 0x43,
-        0x6d, 0xeb, 0x2b, 0x42, 0x6d, 0xed, 0x33, 0x41, 0x6d, 0xfd, 0x7b, 0x49, 0xa9, 0xf4, 0x4f,
-        0x48, 0xa9, 0xf6, 0x57, 0x47, 0xa9, 0xf8, 0x5f, 0x46, 0xa9, 0xfa, 0x67, 0x45, 0xa9, 0xfc,
-        0x6f, 0x44, 0xa9, 0xef, 0x3b, 0xca, 0x6c, 0xc0, 0x03, 0x5f, 0xd6,
-    ],
-    start_offset: 4,
-    location_name: "display_css",
     cached_offset: OnceCell::new(),
 };
 
@@ -178,20 +105,6 @@ pub static LOC_SET_ROOM_ID: SSBUMemoryLocation = SSBUMemoryLocation {
     cached_offset: OnceCell::new(),
 };
 
-pub static LOC_UPDATE_CSS: SSBUMemoryLocation = SSBUMemoryLocation {
-    signature: &[
-        0xea, 0x0f, 0x18, 0xfc, 0xe9, 0x23, 0x01, 0x6d, 0xfc, 0x6f, 0x02, 0xa9, 0xfa, 0x67, 0x03,
-        0xa9, 0xf8, 0x5f, 0x04, 0xa9, 0xf6, 0x57, 0x05, 0xa9, 0xf4, 0x4f, 0x06, 0xa9, 0xfd, 0x7b,
-        0x07, 0xa9, 0xfd, 0xc3, 0x01, 0x91, 0xff, 0x83, 0x0b, 0xd1, 0x08, 0x3c, 0x41, 0xb9, 0xf3,
-        0x03, 0x00, 0xaa, 0x08, 0x01, 0x00, 0x35, 0x68, 0xce, 0x40, 0xf9, 0x08, 0x01, 0x40, 0xf9,
-        0x00, 0x01, 0x40, 0xf9, 0xe1, 0x03, 0x00, 0x32, 0xef, 0x89, 0x75, 0x94, 0x60, 0xfa, 0x46,
-        0xf9, 0x1d, 0xf2, 0x62, 0x94,
-    ],
-    start_offset: 0,
-    location_name: "update_css",
-    cached_offset: OnceCell::new(),
-};
-
 pub static LOC_UPDATE_ROOM: SSBUMemoryLocation = SSBUMemoryLocation {
     signature: &[
         0xff, 0x03, 0x01, 0xd1, 0xf6, 0x57, 0x01, 0xa9, 0xf4, 0x4f, 0x02, 0xa9, 0xfd, 0x7b, 0x03,
@@ -205,6 +118,100 @@ pub static LOC_UPDATE_ROOM: SSBUMemoryLocation = SSBUMemoryLocation {
     location_name: "update_room",
     cached_offset: OnceCell::new(),
 };
+
+#[cfg(feature = "css_ui")]
+pub mod css_ui {
+    use once_cell::sync::OnceCell;
+
+    use super::SSBUMemoryLocation;
+
+    pub static LOC_DRAW: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0x08, 0x0c, 0x40, 0xf9, 0xc8, 0x03, 0x00, 0xb4, 0xff, 0x83, 0x01, 0xd1, 0xf5, 0x1b,
+            0x00, 0xf9, 0xf4, 0x4f, 0x04, 0xa9, 0xfd, 0x7b, 0x05, 0xa9, 0xfd, 0x43, 0x01, 0x91,
+            0xf4, 0x03, 0x00, 0xaa,
+        ],
+        start_offset: 0,
+        location_name: "draw",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_INGAME_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0xff, 0x83, 0x01, 0xd1, 0xfa, 0x67, 0x01, 0xa9, 0xf8, 0x5f, 0x02, 0xa9, 0xf6, 0x57,
+            0x03, 0xa9, 0xf4, 0x4f, 0x04, 0xa9, 0xfd, 0x7b, 0x05, 0xa9, 0xfd, 0x43, 0x01, 0x91,
+            0xf5, 0x03, 0x01, 0xaa, 0xf3, 0x03, 0x00, 0xaa, 0x16, 0x40, 0x01, 0x91, 0x01, 0x37,
+            0x80, 0x52, 0xe0, 0x03, 0x1c, 0x32,
+        ],
+        start_offset: 48,
+        location_name: "ingame_scene",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_ONLINE_MELEE_ANY_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0xe8, 0x69, 0x01, 0xb0, 0x08, 0x01, 0x00, 0x91, 0x1f, 0x40, 0x00, 0x39, 0x08, 0x7c,
+            0x00, 0xa9, 0x1f, 0x40, 0x01, 0x39, 0x1f, 0x54, 0x00, 0xb9, 0x1f, 0x2c, 0x00, 0xf9,
+            0xfd, 0x7b, 0x41, 0xa9, 0xff, 0x83, 0x00, 0x91, 0xc0, 0x03, 0x5f, 0xd6,
+        ],
+        start_offset: 12,
+        location_name: "online_melee_any_scene",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_MAIN_MENU_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0x48, 0x6a, 0x01, 0xb0, 0x08, 0xe1, 0x18, 0x91, 0x7f, 0x42, 0x00, 0x39, 0x68, 0x7e,
+            0x00, 0xa9, 0xa8, 0x7e, 0x01, 0xb0, 0x7f, 0x42, 0x01, 0x39, 0xe1, 0x03, 0x00, 0x32,
+            0x7f, 0x56, 0x00, 0xb9, 0x7f, 0x2e, 0x00, 0xf9, 0x00, 0xf9, 0x43, 0xf9, 0xc7, 0x43,
+            0x3e, 0x94, 0xe0, 0x03, 0x13, 0xaa, 0xfd, 0x7b, 0x42, 0xa9, 0xf3, 0x0b, 0x40, 0xf9,
+            0xff, 0xc3, 0x00, 0x91, 0xc0, 0x03, 0x5f, 0xd6,
+        ],
+        start_offset: 12,
+        location_name: "main_menu_scene",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_MELEE_NORMAL_SEQUENCE_SCENE: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0xc8, 0x6e, 0x01, 0xf0, 0x08, 0xc1, 0x0b, 0x91, 0x60, 0x42, 0x01, 0x91, 0x7f, 0x42,
+            0x00, 0x39, 0x68, 0x7e, 0x00, 0xa9, 0x64, 0x1f, 0x51, 0x94, 0x28, 0x6e, 0x01, 0xd0,
+            0x08, 0xc1, 0x01, 0x91, 0xe9, 0x6e, 0x01, 0x90, 0x29, 0xe1, 0x29, 0x91, 0x7f, 0xc2,
+            0x01, 0x79, 0x7f, 0xe6, 0x00, 0xb9, 0x68, 0x2a, 0x00, 0xf9, 0x69, 0x02, 0x00, 0xf9,
+            0xe0, 0x03, 0x13, 0xaa, 0x7f, 0x76, 0x00, 0xf9, 0xfd, 0x7b, 0x42, 0xa9, 0xf3, 0x0b,
+            0x40, 0xf9, 0xff, 0xc3, 0x00, 0x91, 0xc0, 0x03, 0x5f, 0xd6,
+        ],
+        start_offset: 20,
+        location_name: "melee_normal_sequence_scene",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_DISPLAY_CSS: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0xa0, 0x83, 0x53, 0xf8, 0x34, 0x00, 0x00, 0x94, 0xff, 0x03, 0x0e, 0x91, 0xe9, 0x23,
+            0x43, 0x6d, 0xeb, 0x2b, 0x42, 0x6d, 0xed, 0x33, 0x41, 0x6d, 0xfd, 0x7b, 0x49, 0xa9,
+            0xf4, 0x4f, 0x48, 0xa9, 0xf6, 0x57, 0x47, 0xa9, 0xf8, 0x5f, 0x46, 0xa9, 0xfa, 0x67,
+            0x45, 0xa9, 0xfc, 0x6f, 0x44, 0xa9, 0xef, 0x3b, 0xca, 0x6c, 0xc0, 0x03, 0x5f, 0xd6,
+        ],
+        start_offset: 4,
+        location_name: "display_css",
+        cached_offset: OnceCell::new(),
+    };
+
+    pub static LOC_UPDATE_CSS: SSBUMemoryLocation = SSBUMemoryLocation {
+        signature: &[
+            0xea, 0x0f, 0x18, 0xfc, 0xe9, 0x23, 0x01, 0x6d, 0xfc, 0x6f, 0x02, 0xa9, 0xfa, 0x67,
+            0x03, 0xa9, 0xf8, 0x5f, 0x04, 0xa9, 0xf6, 0x57, 0x05, 0xa9, 0xf4, 0x4f, 0x06, 0xa9,
+            0xfd, 0x7b, 0x07, 0xa9, 0xfd, 0xc3, 0x01, 0x91, 0xff, 0x83, 0x0b, 0xd1, 0x08, 0x3c,
+            0x41, 0xb9, 0xf3, 0x03, 0x00, 0xaa, 0x08, 0x01, 0x00, 0x35, 0x68, 0xce, 0x40, 0xf9,
+            0x08, 0x01, 0x40, 0xf9, 0x00, 0x01, 0x40, 0xf9, 0xe1, 0x03, 0x00, 0x32, 0xef, 0x89,
+            0x75, 0x94, 0x60, 0xfa, 0x46, 0xf9, 0x1d, 0xf2, 0x62, 0x94,
+        ],
+        start_offset: 0,
+        location_name: "update_css",
+        cached_offset: OnceCell::new(),
+    };
+}
 
 // TODO: not findable atm, probably need to change search algo.
 // this and `find_pane_by_name_recursive` are currently the only hardcoded offsets
